SMILE/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
groq==0.11.0
grpcio==1.65.4
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
inspector==0.5.1.1
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
moviepy==1.0.3
msgpack==1.1.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.45.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
protobuf==4.25.4
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.7.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.5.0.post1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow_keras==0.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
waitress==3.0.0
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
